changed to fully comply with erc20

+++ contracts/DigitizeCoin.sol
@@ -37,7 +37,7 @@ contract DigitizeCoin is Ownable {
 
   string public constant name = "Digitize Coin";
   string public constant symbol = "DTZ";
-  uint256 public constant decimals = 18;
+  uint8 public constant decimals = 18;
   uint256 public constant initialSupply = 200000000 * (10 ** decimals);
   uint256 public totalSupply;
